refactor(web): split api.ts into per-domain modules; unify SSE auth path (#1763)
web/src/lib/api.ts (2,809 lines: ~165 interfaces + a ~170-method object
literal) is now a 62-line barrel over 17 domain modules in
web/src/lib/api/:

  core.ts            128  fetchApi, ApiError, 401/409 login redirect
  shared.ts           36  DeviceType/PageType/LineMetadata/RowConfig/ActionResponse
  system.ts          331  status/start/stop, version, self-update, debug, MQTT, backup, WiFi
  boards.ts          356  board instances/settings, FiestaPanel, board config
  setup.ts            81  setup wizard (connection test, Local API, scan, welcome)
  pages.ts           245  page CRUD/preview/share, staff picks, active page
  templates.ts       119  variables, formula functions, validate/render, HA entities
  schedules.ts       301  schedule CRUD/validation, overrides, silence
  collections.ts      91  collections CRUD
  settings.ts        195  general/polling/display/location/beta/plugin settings
  transitions.ts     106  transition plugins (beta)
  plugins.ts         299  installed plugin system, instances, demo pages
  plugin-registry.ts 102  marketplace, git install, plugin updates
  ai.ts              102  BYO-LLM settings + one-shot generation
  auth.ts            190  session/auth helpers, MCP bearer token (#1825 preserved)
  misc.ts            230  legacy displays + helper endpoints (Muni, Bay Wheels, ...)

Zero call-site churn: the barrel re-exports every type and composes the
same `api` object from the domain objects, so `import { api, type X }
from "@/lib/api"` is unchanged everywhere. Verified: `grep -rn 'from
"@/lib/api"' web/src | wc -l` is 145 before and after (web/app: 14 both
sides); a key-set diff of the old object literal vs the composed object
shows the identical 178 methods with no duplicates; an export-name diff
shows no symbol lost. Bodies were sliced from the original file
line-for-line, not retyped. tsc --noEmit clean; full vitest suite 1611
passed / 7 skipped before and after (plus the 2 new tests below).

SSE auth unification: lib/api-stream.ts (the /pages/ai/chat stream) used
to bypass the login redirect entirely. Its onopen error path now calls
the shared redirectToLoginIfNeeded() exported from api/core.ts, so an
expired-session chat lands on /login like every fetchApi call, with
streaming semantics untouched. Fail-first: the new api-stream test
"routes a 401 response through the shared login redirect" was written
and run before wiring — it failed with `expected "vi.fn()" to be called
with arguments … Number of calls: 0` — then passed after the one-line
call was added. A companion test pins that a plain 500 still does not
redirect.

Drift fix: `demo_plugin_id?: string | null` added to the TS `Page` and
`PageCreate` interfaces, mirroring `demo_plugin_id: str | None = None`
in src/pages/models.py (PageUpdate has no such field in Python, so TS
stays without it too).

Drift scan (Page/PageCreate/Schedule/Collection/BoardStatus vs
src/*/models.py): no other missing fields found. Two deliberate
asymmetries noted, not "fixed": TS ScheduleEntry's resolved_start_time /
resolved_end_time are added by the route layer (api_server.py) rather
than the Pydantic model, and TS BoardStatus matches the dict built
inline in the /status route (no Pydantic model exists for it).

Closes #1763.
